refactor(renderer): extract browser results controller
